Refactor layout and column services to enhance LaTeX document generation; update data handling in LayoutColumn and BackgroungImage classes.

+++ backend/src/main/java/com/coigniez/resumebuilder/domain/columnholder/header/Header.java
@@ -15,7 +15,7 @@ import lombok.experimental.SuperBuilder;
 public class Header extends ColumnHolder {
 
     @Column(precision = 4)
-    private Double height;
+    private Float height;
     private Boolean repeatOnEveryPage;
 
     @OneToOne
